fix: mark auto-minimize-policy comments as seen to prevent double-surfacing
Bot comments matched by the minimize policy were added to `actionable`
and `minimizeIds` but never to `toMarkSeen`, so no seen marker was
written. After `resolve --minimize-comment-ids` minimized them on
GitHub, the next fetch saw them as `isMinimized: true`, found no marker,
and re-surfaced them as first-look items with `[status: minimized]`.

Fixes the duplicate display observed in jonathanong/filaments#4245
(Gemini quota warning and CodeRabbit rate-limit comments shown twice).
