+++ src/commands/check.runcheck-first-look-items.mock.test.mts
@@ -2,15 +2,12 @@ import { describe, it, expect } from "vitest";
 import {
   registerHooks,
   BASE_OPTS,
-  defaultConfig,
   makeBatchData,
   mockAutoResolveOutdated,
   mockFetchPrBatch,
-  mockLoadConfig,
   mockLoadSeenMap,
   mockMarkSeen,
   makeThread,
-  makeComment,
 } from "../../test-helpers/commands/check.test-support.mts";
 import { hashBody } from "../state/seen-comments.mts";
 import { runCheck } from "./check.mts";
@@ -21,7 +18,6 @@ describe("runCheck — first-look items", () => {
   it("surfaces unseen outdated thread in threads.firstLook", async () => {
     const outdated = makeThread({ id: "t-outdated", isOutdated: true });
     mockFetchPrBatch.mockResolvedValue({ data: makeBatchData({ reviewThreads: [outdated] }) });
-    mockLoadSeenMap.mockResolvedValue(new Map());
     const report = await runCheck(BASE_OPTS);
     expect(report.threads.firstLook).toHaveLength(1);
     expect(report.threads.firstLook[0]?.id).toBe("t-outdated");
@@ -31,7 +27,6 @@ describe("runCheck — first-look items", () => {
   it("surfaces unseen active thread in threads.actionable and marks it seen", async () => {
     const active = makeThread({ id: "t-active", body: "active feedback" });
     mockFetchPrBatch.mockResolvedValue({ data: makeBatchData({ reviewThreads: [active] }) });
-    mockLoadSeenMap.mockResolvedValue(new Map());
 
     const report = await runCheck(BASE_OPTS);
 
@@ -68,7 +63,6 @@ describe("runCheck — first-look items", () => {
   it("does not auto-resolve outdated threads during check", async () => {
     const outdated = makeThread({ id: "t-auto", isOutdated: true });
     mockFetchPrBatch.mockResolvedValue({ data: makeBatchData({ reviewThreads: [outdated] }) });
-    mockLoadSeenMap.mockResolvedValue(new Map());
     mockAutoResolveOutdated.mockResolvedValue({ resolved: ["t-auto"], errors: [] });
     const report = await runCheck({ ...BASE_OPTS, autoResolve: true });
     expect(mockAutoResolveOutdated).not.toHaveBeenCalled();
@@ -78,7 +72,6 @@ describe("runCheck — first-look items", () => {
   it("surfaces unseen resolved thread in threads.firstLook", async () => {
     const resolved = makeThread({ id: "t-resolved", isResolved: true });
     mockFetchPrBatch.mockResolvedValue({ data: makeBatchData({ reviewThreads: [resolved] }) });
-    mockLoadSeenMap.mockResolvedValue(new Map());
     const report = await runCheck(BASE_OPTS);
     expect(report.threads.firstLook).toHaveLength(1);
     expect(report.threads.firstLook[0]?.firstLookStatus).toBe("resolved");
@@ -86,84 +79,9 @@ describe("runCheck — first-look items", () => {
   it("surfaces unseen minimized thread in threads.firstLook", async () => {
     const minimized = makeThread({ id: "t-minimized", isMinimized: true });
     mockFetchPrBatch.mockResolvedValue({ data: makeBatchData({ reviewThreads: [minimized] }) });
-    mockLoadSeenMap.mockResolvedValue(new Map());
     const report = await runCheck(BASE_OPTS);
     expect(report.threads.firstLook).toHaveLength(1);
     expect(report.threads.firstLook[0]?.firstLookStatus).toBe("minimized");
-  });
-  it("surfaces unseen minimized comment in comments.firstLook", async () => {
-    const minimized = makeComment({ id: "c-min", isMinimized: true });
-    mockFetchPrBatch.mockResolvedValue({ data: makeBatchData({ comments: [minimized] }) });
-    mockLoadSeenMap.mockResolvedValue(new Map());
-
-    const report = await runCheck(BASE_OPTS);
-    expect(report.comments.firstLook).toHaveLength(1);
-    expect(report.comments.firstLook[0]?.firstLookStatus).toBe("minimized");
-  });
-  it("marker-gates visible comments excluded by minimizeComments policy", async () => {
-    const cfg = defaultConfig();
-    cfg.iterate.minimizeComments = "bots";
-    mockLoadConfig.mockReturnValue(cfg);
-    const human = makeComment({
-      id: "c-human",
-      author: "alice",
-      authorType: "User",
-      body: "please consider this",
-    });
-    const bot = makeComment({
-      id: "c-bot",
-      author: "app",
-      authorType: "Bot",
-      body: "automated note",
-    });
-    mockFetchPrBatch.mockResolvedValue({ data: makeBatchData({ comments: [human, bot] }) });
-    mockLoadSeenMap.mockResolvedValue(new Map());
-
-    const report = await runCheck(BASE_OPTS);
-
-    expect(report.comments.actionable.map((c) => c.id)).toEqual(["c-human", "c-bot"]);
-    expect(report.comments.minimizeIds).toEqual(["c-bot"]);
-    expect(mockMarkSeen).toHaveBeenCalledWith(
-      expect.any(Object),
-      "c-human",
-      "please consider this",
-    );
-    expect(mockMarkSeen).not.toHaveBeenCalledWith(expect.any(Object), "c-bot", expect.anything());
-  });
-  it("suppresses unchanged visible comments excluded by minimizeComments policy", async () => {
-    const cfg = defaultConfig();
-    cfg.iterate.minimizeComments = "bots";
-    mockLoadConfig.mockReturnValue(cfg);
-    const human = makeComment({
-      id: "c-human",
-      authorType: "User",
-      body: "already seen",
-    });
-    mockFetchPrBatch.mockResolvedValue({ data: makeBatchData({ comments: [human] }) });
-    mockLoadSeenMap.mockResolvedValue(
-      new Map([["c-human", { seenAt: 1000, bodyHash: hashBody("already seen") }]]),
-    );
-
-    const report = await runCheck(BASE_OPTS);
-
-    expect(report.comments.actionable).toEqual([]);
-    expect(report.comments.minimizeIds).toEqual([]);
-  });
-  it("re-surfaces edited visible comments excluded by minimizeComments policy", async () => {
-    const cfg = defaultConfig();
-    cfg.iterate.minimizeComments = "none";
-    mockLoadConfig.mockReturnValue(cfg);
-    const comment = makeComment({ id: "c-human", authorType: "User", body: "new text" });
-    mockFetchPrBatch.mockResolvedValue({ data: makeBatchData({ comments: [comment] }) });
-    mockLoadSeenMap.mockResolvedValue(
-      new Map([["c-human", { seenAt: 1000, bodyHash: hashBody("old text") }]]),
-    );
-
-    const report = await runCheck(BASE_OPTS);
-
-    expect(report.comments.actionable).toMatchObject([{ id: "c-human", edited: true }]);
-    expect(report.comments.minimizeIds).toEqual([]);
-    expect(mockMarkSeen).toHaveBeenCalledWith(expect.any(Object), "c-human", "new text");
   });
   it("suppresses already-seen items (unchanged hash)", async () => {
     const outdated = makeThread({ id: "t-outdated", isOutdated: true, body: "fix this" });
